feat(compiler) Created abstraction for Reg and Location
