+++ lib/compiler-singlepass/src/x64_decl.rs
@@ -1,6 +1,7 @@
 //! X64 structures.
 
 use crate::common_decl::{MachineState, MachineValue, RegisterIndex};
+use crate::location::Reg as AbstractReg;
 use std::collections::BTreeMap;
 use wasmer_compiler::CallingConvention;
 use wasmer_types::Type;
@@ -9,38 +10,22 @@ use wasmer_types::Type;
 #[repr(u8)]
 #[derive(Copy, Clone, Debug, Eq, PartialEq, Ord, PartialOrd, Hash)]
 pub enum GPR {
-    /// RAX register
-    RAX,
-    /// RCX register
-    RCX,
-    /// RDX register
-    RDX,
-    /// RBX register
-    RBX,
-    /// RSP register
-    RSP,
-    /// RBP register
-    RBP,
-    /// RSI register
-    RSI,
-    /// RDI register
-    RDI,
-    /// R8 register
-    R8,
-    /// R9 register
-    R9,
-    /// R10 register
-    R10,
-    /// R11 register
-    R11,
-    /// R12 register
-    R12,
-    /// R13 register
-    R13,
-    /// R14 register
-    R14,
-    /// R15 register
-    R15,
+    RAX = 0,
+    RCX = 1,
+    RDX = 2,
+    RBX = 3,
+    RSP = 4,
+    RBP = 5,
+    RSI = 6,
+    RDI = 7,
+    R8 = 8,
+    R9 = 9,
+    R10 = 10,
+    R11 = 11,
+    R12 = 12,
+    R13 = 13,
+    R14 = 14,
+    R15 = 15,
 }
 
 /// XMM registers.
@@ -48,38 +33,102 @@ pub enum GPR {
 #[derive(Copy, Clone, Debug, Eq, PartialEq, Ord, PartialOrd, Hash)]
 #[allow(dead_code)]
 pub enum XMM {
-    /// XMM register 0
-    XMM0,
-    /// XMM register 1
-    XMM1,
-    /// XMM register 2
-    XMM2,
-    /// XMM register 3
-    XMM3,
-    /// XMM register 4
-    XMM4,
-    /// XMM register 5
-    XMM5,
-    /// XMM register 6
-    XMM6,
-    /// XMM register 7
-    XMM7,
-    /// XMM register 8
-    XMM8,
-    /// XMM register 9
-    XMM9,
-    /// XMM register 10
-    XMM10,
-    /// XMM register 11
-    XMM11,
-    /// XMM register 12
-    XMM12,
-    /// XMM register 13
-    XMM13,
-    /// XMM register 14
-    XMM14,
-    /// XMM register 15
-    XMM15,
+    XMM0 = 0,
+    XMM1 = 1,
+    XMM2 = 2,
+    XMM3 = 3,
+    XMM4 = 4,
+    XMM5 = 5,
+    XMM6 = 6,
+    XMM7 = 7,
+    XMM8 = 8,
+    XMM9 = 9,
+    XMM10 = 10,
+    XMM11 = 11,
+    XMM12 = 12,
+    XMM13 = 13,
+    XMM14 = 14,
+    XMM15 = 15,
+}
+
+impl AbstractReg for GPR {
+    fn is_callee_save(self) -> bool {
+        const IS_CALLEE_SAVE: [bool; 16] = [
+            false, false, false, true, true, true, false, false, false, false, false, false, true,
+            true, true, true,
+        ];
+        IS_CALLEE_SAVE[self as usize]
+    }
+    fn is_reserved(self) -> bool {
+        self == GPR::RSP || self == GPR::RBP || self == GPR::R10 || self == GPR::R15
+    }
+    fn into_index(self) -> usize {
+        self as usize
+    }
+    fn from_index(n: usize) -> Result<GPR, ()> {
+        const REGS: [GPR; 16] = [
+            GPR::RAX,
+            GPR::RCX,
+            GPR::RDX,
+            GPR::RBX,
+            GPR::RSP,
+            GPR::RBP,
+            GPR::RSI,
+            GPR::RDI,
+            GPR::R8,
+            GPR::R9,
+            GPR::R10,
+            GPR::R11,
+            GPR::R12,
+            GPR::R13,
+            GPR::R14,
+            GPR::R15,
+        ];
+        match n {
+            0..=15 => Ok(REGS[n]),
+            _ => Err(()),
+        }
+    }
+}
+
+impl AbstractReg for XMM {
+    fn is_callee_save(self) -> bool {
+        const IS_CALLEE_SAVE: [bool; 16] = [
+            false, false, false, false, false, false, false, false, 
+            true, true, true, true, true, true, true, true,
+        ];
+        IS_CALLEE_SAVE[self as usize]
+    }
+    fn is_reserved(self) -> bool {
+        false
+    }
+    fn into_index(self) -> usize {
+        self as usize
+    }
+    fn from_index(n: usize) -> Result<XMM, ()> {
+        const REGS: [XMM; 16] = [
+            XMM::XMM0,
+            XMM::XMM1,
+            XMM::XMM2,
+            XMM::XMM3,
+            XMM::XMM4,
+            XMM::XMM5,
+            XMM::XMM6,
+            XMM::XMM7,
+            XMM::XMM8,
+            XMM::XMM9,
+            XMM::XMM10,
+            XMM::XMM11,
+            XMM::XMM12,
+            XMM::XMM13,
+            XMM::XMM14,
+            XMM::XMM15,
+        ];
+        match n {
+            0..=15 => Ok(REGS[n]),
+            _ => Err(()),
+        }
+    }
 }
 
 /// A machine register under the x86-64 architecture.
@@ -103,31 +152,8 @@ impl X64Register {
     /// Converts a DWARF regnum to X64Register.
     pub fn _from_dwarf_regnum(x: u16) -> Option<X64Register> {
         Some(match x {
-            0 => X64Register::GPR(GPR::RAX),
-            1 => X64Register::GPR(GPR::RDX),
-            2 => X64Register::GPR(GPR::RCX),
-            3 => X64Register::GPR(GPR::RBX),
-            4 => X64Register::GPR(GPR::RSI),
-            5 => X64Register::GPR(GPR::RDI),
-            6 => X64Register::GPR(GPR::RBP),
-            7 => X64Register::GPR(GPR::RSP),
-            8 => X64Register::GPR(GPR::R8),
-            9 => X64Register::GPR(GPR::R9),
-            10 => X64Register::GPR(GPR::R10),
-            11 => X64Register::GPR(GPR::R11),
-            12 => X64Register::GPR(GPR::R12),
-            13 => X64Register::GPR(GPR::R13),
-            14 => X64Register::GPR(GPR::R14),
-            15 => X64Register::GPR(GPR::R15),
-
-            17 => X64Register::XMM(XMM::XMM0),
-            18 => X64Register::XMM(XMM::XMM1),
-            19 => X64Register::XMM(XMM::XMM2),
-            20 => X64Register::XMM(XMM::XMM3),
-            21 => X64Register::XMM(XMM::XMM4),
-            22 => X64Register::XMM(XMM::XMM5),
-            23 => X64Register::XMM(XMM::XMM6),
-            24 => X64Register::XMM(XMM::XMM7),
+            0..=15 => X64Register::GPR(GPR::from_index(x as usize).unwrap()),
+            17..=24 => X64Register::XMM(XMM::from_index(x as usize - 17).unwrap()),
             _ => return None,
         })
     }
